refactor: move validate response function
