+++ src/components/probe/prober/index.test.ts
@@ -23,268 +23,11 @@
  **********************************************************************************/
 
 import { expect } from '@oclif/test'
-import {
-  probeRequestResult,
-  type ProbeRequestResponse,
-} from '../../../interfaces/request'
 import type { Probe } from '../../../interfaces/probe'
 import { createProber } from '../../../components/probe/prober/factory'
 import type { ServerAlertState } from '../../../interfaces/probe-status'
 import type { ValidatedResponse } from '../../../plugins/validate-response'
 import { serverAlertStateInterpreters } from '../../notification/process-server-status'
-
-describe('validateResponse', () => {
-  const mockedAlerts = [
-    {
-      assertion: 'response.status < 200 or response.status > 299',
-      message: '',
-    },
-    { assertion: 'response.time > 10', message: '' },
-  ]
-
-  it('status-not-2xx = true && response-time-greater-than-10-ms = true', () => {
-    const prober = createProber({
-      counter: 0,
-      notifications: [],
-      probeConfig: { alerts: mockedAlerts } as Probe,
-    })
-    const res = generateMockedResponse(300, 20, true)
-    const data = prober.validateResponse(res)
-
-    expect(data).to.eql([
-      {
-        alert: {
-          assertion: 'response.status < 200 or response.status > 299',
-          message: '',
-        },
-        isAlertTriggered: true,
-        response: {
-          data: '',
-          body: '',
-          responseTime: 20,
-          headers: {},
-          status: 300,
-          result: probeRequestResult.success,
-          isProbeResponsive: true,
-        },
-      },
-      {
-        alert: {
-          assertion: 'response.time > 10',
-          message: '',
-        },
-        response: {
-          data: '',
-          body: '',
-          responseTime: 20,
-          headers: {},
-          status: 300,
-          result: probeRequestResult.success,
-          isProbeResponsive: true,
-        },
-        isAlertTriggered: true,
-      },
-    ])
-  })
-
-  it('status-not-2xx = false && response-time-greater-than-10-ms = true', () => {
-    const prober = createProber({
-      counter: 0,
-      notifications: [],
-      probeConfig: { alerts: mockedAlerts } as Probe,
-    })
-    const res = generateMockedResponse(200, 20, true)
-    const data = prober.validateResponse(res)
-
-    expect(data).to.eql([
-      {
-        alert: {
-          assertion: 'response.status < 200 or response.status > 299',
-          message: '',
-        },
-        response: {
-          data: '',
-          body: '',
-          responseTime: 20,
-          headers: {},
-          status: 200,
-          result: probeRequestResult.success,
-          isProbeResponsive: true,
-        },
-        isAlertTriggered: false,
-      },
-      {
-        alert: {
-          assertion: 'response.time > 10',
-          message: '',
-        },
-        response: {
-          data: '',
-          body: '',
-          responseTime: 20,
-          headers: {},
-          status: 200,
-          result: probeRequestResult.success,
-          isProbeResponsive: true,
-        },
-        isAlertTriggered: true,
-      },
-    ])
-  })
-
-  it('status-not-2xx = true && response-time-greater-than-10-ms = false', () => {
-    const prober = createProber({
-      counter: 0,
-      notifications: [],
-      probeConfig: { alerts: mockedAlerts } as Probe,
-    })
-    const res = generateMockedResponse(300, 10, true)
-    const data = prober.validateResponse(res)
-
-    expect(data).to.eql([
-      {
-        alert: {
-          assertion: 'response.status < 200 or response.status > 299',
-          message: '',
-        },
-        response: {
-          data: '',
-          body: '',
-          responseTime: 10,
-          headers: {},
-          status: 300,
-          result: probeRequestResult.success,
-          isProbeResponsive: true,
-        },
-        isAlertTriggered: true,
-      },
-      {
-        alert: {
-          assertion: 'response.time > 10',
-          message: '',
-        },
-        response: {
-          data: '',
-          body: '',
-          responseTime: 10,
-          headers: {},
-          status: 300,
-          result: probeRequestResult.success,
-          isProbeResponsive: true,
-        },
-        isAlertTriggered: false,
-      },
-    ])
-  })
-
-  it('status-not-2xx = false && response-time-greater-than-10-ms = false', () => {
-    const prober = createProber({
-      counter: 0,
-      notifications: [],
-      probeConfig: { alerts: mockedAlerts } as Probe,
-    })
-    const res = generateMockedResponse(200, 10, true)
-    const data = prober.validateResponse(res)
-
-    expect(data).to.eql([
-      {
-        alert: {
-          assertion: 'response.status < 200 or response.status > 299',
-          message: '',
-        },
-        response: {
-          data: '',
-          body: '',
-          responseTime: 10,
-          headers: {},
-          status: 200,
-          result: probeRequestResult.success,
-          isProbeResponsive: true,
-        },
-        isAlertTriggered: false,
-      },
-      {
-        alert: {
-          assertion: 'response.time > 10',
-          message: '',
-        },
-        response: {
-          data: '',
-          body: '',
-          responseTime: 10,
-          headers: {},
-          status: 200,
-          result: probeRequestResult.success,
-          isProbeResponsive: true,
-        },
-        isAlertTriggered: false,
-      },
-    ])
-  })
-
-  it('should assert with additional query', () => {
-    const prober = createProber({
-      counter: 0,
-      notifications: [],
-      probeConfig: { alerts: mockedAlerts } as Probe,
-    })
-    const res = generateMockedResponse(200, 10, true)
-    const data = prober.validateResponse(res, [
-      { assertion: 'response.time > 5', message: '' },
-    ])
-
-    expect(data).to.eql([
-      {
-        alert: {
-          assertion: 'response.status < 200 or response.status > 299',
-          message: '',
-        },
-        response: {
-          data: '',
-          body: '',
-          responseTime: 10,
-          headers: {},
-          status: 200,
-          result: probeRequestResult.success,
-          isProbeResponsive: true,
-        },
-        isAlertTriggered: false,
-      },
-      {
-        alert: {
-          assertion: 'response.time > 10',
-          message: '',
-        },
-        response: {
-          data: '',
-          body: '',
-          responseTime: 10,
-          headers: {},
-          status: 200,
-          result: probeRequestResult.success,
-          isProbeResponsive: true,
-        },
-        isAlertTriggered: false,
-      },
-      {
-        alert: {
-          assertion: 'response.time > 5',
-          message: '',
-        },
-        response: {
-          data: '',
-          body: '',
-          responseTime: 10,
-          headers: {},
-          status: 200,
-          result: probeRequestResult.success,
-          isProbeResponsive: true,
-        },
-        isAlertTriggered: true,
-      },
-    ])
-  })
-})
 
 describe('processThresholds', () => {
   beforeEach(() => {
@@ -449,21 +192,3 @@ describe('processThresholds', () => {
     expect(result[0].state).eq('UP')
   })
 })
-
-function generateMockedResponse(
-  status: number,
-  responseTime: number,
-  isProbeResponsive: boolean
-): ProbeRequestResponse {
-  return {
-    data: '',
-    body: '',
-    status,
-    responseTime,
-    headers: {},
-    result: isProbeResponsive
-      ? probeRequestResult.success
-      : probeRequestResult.failed,
-    isProbeResponsive,
-  }
-}
